Correct 50 typos in the text.
These errors were identified by comparing the text against the text of
a Canadian printing, filtering out differences due to regional
variations (e.g. spelling and hyphenation), and then manually
examining the remaining differences to identify legitimate mistakes in
the text.

(The only alteration whose rationale might be unclear is the final
one in chapter 23, in which the quotation mark at the end of a
paragraph is removed. A careful reading of the subsequent text,
however, will make it clear that the following paragraph is a
continuation of Detective Parker's speech, and that Wimsey does not
speak until the paragraph after that.)